refactor: StationSummaryWindow -> StationInfoWindow 으로 이름 변경
[#820]
