+++ frontend/src/components/ui/StationInfoWindow/SummaryButtons.tsx
