+++ frontend/src/components/ui/StationInfoWindow/index.tsx
@@ -1,0 +1,3 @@
+import StationInfoWindow from './StationInfoWindow';
+
+export default StationInfoWindow;
