+++ frontend/src/components/ui/StationInfoWindow/useFetchStationDetatils.ts
