--- frontend/src/components/ui/StationSummaryWindow/index.tsx
@@ -1,3 +1,0 @@
-import StationSummaryWindow from './StationSummaryWindow';
-
-export default StationSummaryWindow;
